+++ LibraryApi/DTOs/Book/BookViewDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Title { get; set; }
-        public uint Price { get; set; }
+        public decimal Price { get; set; }
         public string AuthorName { get; set; }
         public string GenreName { get; set; }
     }
