+++ LibraryApi/Repository/Specific/IBookRepository.cs
@@ -8,6 +8,6 @@ namespace LibraryApi.Repository.Specific
         Task<Book> GetByIdIncludingAsync(int id);
         Task<IEnumerable<Book>> GetBooksByAuthorNameAsync(string authorName);
         Task<IEnumerable<Book>> GetBooksByGenreNameAsync(string genreName);
-        Task<IEnumerable<Book>> GetBooksByPriceRangeAsync(uint Min, uint Max);
+        Task<IEnumerable<Book>> GetBooksByPriceRangeAsync(decimal Min, decimal Max);
     }
 }
